config.py
"""Configuration settings for the blog agent system."""

from google.genai import types

# Configure retry options for LLM calls
retry_config = types.HttpRetryOptions(
    attempts=5,
    exp_base=7,
    initial_delay=1,
    http_status_codes=[429, 500, 503, 504],
)

